Separating out Mezzurite plugin export and implementation
